+++ src/main/java/gin/melec/SplitLeft.java
@@ -27,7 +27,7 @@ public class SplitLeft extends AbstractSplit {
      * Public constructor for a split with a position.
      * @param position , the position of the border.
      */
-    public SplitLeft(final int position) {
+    public SplitLeft(final long position) {
         super(position);
     }
 
@@ -37,12 +37,12 @@ public class SplitLeft extends AbstractSplit {
     }
 
     @Override
-    protected final int xPosition() {
+    protected final long xPosition() {
         return position;
     }
 
     @Override
-    protected final int yPosition() {
+    protected final long yPosition() {
         return 0;
     }
